Fix: Update architecture of the database
Separate the routes and the db config
