+++ back_flask/database.py
@@ -1,0 +1,3 @@
+class Database:
+    SQLALCHEMY_DATABASE_URI = 'postgresql://lucaspesse:root@localhost/my_bookshop'
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
